Create TalleyBookReaderWriter.java
First iteration of the Talley Workbook reader and writer. Is non-functional at the moment since it is unclear how it will be used by the rest of the system. Is able to read in the workbook but doesn't have a specific output. Writer function blank and an additional reader that only reads in the tallies of a single day could be useful.