Vastly Expanded OnCall
Nearly finished but still a work in Progress.

Created 2 functions that will help the main assign function. Created the assign function to parse through the absence record until it finds an uncovered absence. Will then create the appropriate line in the output file and create a new row and restart its parsing.