Updated OnCall
The assign function works now. It creates a row then parses through the absence records. If it finds a teacher who is absent, hasn't been covered yet, and has a course that needs to be covered, then it will assign the most applicable teacher to cover it then it will restart the process.
Still has a couple exceptions that need to be accounted for and has 2 majour flaws. It doesn't update the on-call tally workbook and its selection of a teacher in the workbook is broken.